requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
click==8.1.5
cryptography==3.4.8
dnspython==2.4.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.20.2
httpx==0.24.1
idna==3.4
mysql-connector-python==8.0.33
oauthlib==3.2.0
passlib==1.7.4
protobuf==3.20.3
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
PyJWT==2.3.0
pyparsing==3.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.0
